\name{ar1sim}

\alias{ar1sim}

\title{Simulate AR(1) process.}

\description{Simulates an autoregressive process of order 1.}

\usage{
ar1sim(n = 50, rho, u0 = 0, var.e = 1, details = FALSE, seed = NULL) 
}

\arguments{
  \item{n}{Number of observations to be simulated.}
  \item{u0}{Start value of the process in t = 0.}
  \item{rho}{True rho value of the AR1(1) process to be simulated.}
  \item{var.e}{Variance of the random error. If zero no random error is added.}
  \item{details}{Logical value indicating whether details should be printed.}
  \item{seed}{Optionally set a custom random seed for reproducing results.}
}

\note{Objects generated by ar1sim() can be plotted using the regular \code{plot()} command.

\code{plot.what = "time"} plots simulated AR(1) values over time. Available options are

\tabular{ll}{
\code{...} \tab Other arguments that \code{plot()} understands.\cr
}


\code{plot.what = "lag"} plots simulated AR(1) values over its lagged values. Available options are

\tabular{ll}{
\code{true.line} \tab Logical value (Default: T). Should the true line be plotted?.\cr
\code{acc.line} \tab Logical value (Default: F). Should the autocorrelation coefficient line be plotted?.\cr
\code{ols.line} \tab Logical value (Default: F). Should the ols regression line be plotted?.\cr
\code{...} \tab Other arguments that \code{plot()} understands.\cr
}
}
\value{List object including:

\tabular{ll}{
\code{u.sim} \tab Vector of simulated AR(1) values.\cr
\code{n} \tab Total number of simulated AR(1) values.\cr
\code{rho} \tab True rho value of AR(1) process.\cr
\code{e.sim} \tab Normal errors in AR(1) process.\cr
}
}
\examples{
# Generate 30 positively autocorrelated errors
my.ar1 <- ar1sim(n = 30, rho = 0.9, var.e = 0.1, seed = 511)
my.ar1
plot(my.ar1)

# Illustrate the effect of Rho on the AR(1)
set.seed(12)
par(mfrow = c(2,4), mar = c(1,1,1,1))
rhovalues <- c(0.1, 0.5, 0.8, 0.99)
for (i in c(0, 0.3)){
  for (rho in rhovalues){
    u.data <- ar1sim(n = 20, u0 = 2, rho = rho, var.e = i)
    plot(u.data, plot.what = "lag", cex.legend = 0.7, xlim = c(-2.5,2.5), ylim = c(-2.5,2.5),
         acc.line = TRUE, ols.line = TRUE)
  }
}
par(mfrow = c(1,1))

# Illustrate the effect of Rho on the (non-)stationarity of the AR(1)
set.seed(1324)
par(mfrow = c(2, 4), mar = c(1,1,1,1))
for (rho in c(0.1, 0.9, 1, 1.04, -0.1, -0.9, -1, -1.04)){
  u.data <- ar1sim(n = 25, u0 = 5, rho = rho, var.e = 0)
  plot(u.data, plot.what = "time", ylim = c(-8,8))
}
par(mfrow = c(1,1))
}

\keyword{AR(1) process, autocorrelated errors, simulation}