\name{data.anscombe}
\alias{data.anscombe}
\docType{data}
\title{
Anscombe's four datasets, used in chapter 3 of the workbook.
}
\description{
Four x-y datasets which have the same statistical properties (mean, variance, correlation, regression line, etc.), yet are quite different.
}
\usage{data.anscombe}
\format{
  A data frame of 4 datasets with 11 observations each and x and y variable.
  \describe{
    \item{x1 - x4}{x-variable of the four datasets.}
    \item{y1 - y4}{y-variable of the four datasets.}
  }
}
\details{
These data provide a dataset used in chapter 3 of the workbook.
}
\source{
Tufte, Edward R. (1989) The Visual Display of Quantitative Information, 13-14. Graphics Press.
R package datasets
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
