% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.auto.R
\docType{data}
\name{data.auto}
\alias{data.auto}
\title{Prices and Qualitative Characteristics of US-Cars}
\format{
A data frame with 52 observations on the following nine variables:
\tabular{ll}{
\code{make} \tab make and model.\cr
\code{price} \tab price (in dollar).\cr
\code{mpgall} \tab mileage (miles per gallon).\cr
\code{headroom} \tab headroom (in inch).\cr
\code{trunk} \tab trunk Space (in cubic foot).\cr
\code{weight} \tab weight (in pound).\cr
\code{length} \tab length (in inch).\cr
\code{turn} \tab turn circle (in foot).\cr
\code{displacement} \tab displacement (in cubic inch).\cr
}
}
\source{
Chambers, J.M, W.S. Cleveland, B. Kleiner, P.A. Tukey (1983): Graphical Methods for Data Analysis, Wadsworth International Group.
}
\usage{
data.auto
}
\description{
This is a data set on the prices and qualitative characteristics of US-cars sold in 1979.
}
\details{
In von Auer and Hoffmann (2017, Chap. 13) these data are used to illustrate the selection process of exogenous variables.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
