\name{data.ballb}
\alias{data.ballb}
\docType{data}
\title{
Defective ball bearings (Ch. 16).
}
\description{
Share of defective ball bearings in a factory and the number of shifts between two maintenances.
}
\usage{data.ballb}
\format{
  A data frame with 7 observations on the following 2 variables.
  \describe{
    \item{defbb}{Share of defective ball bearings (thousandth).}
    \item{nshifts}{Number of shifts between two maintenances.}
  }
}
\details{
These data provide a dataset for analyzing the expected value of the error term. Used in chapter 16 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
