% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.burglary.R
\docType{data}
\name{data.burglary}
\alias{data.burglary}
\title{Burglaries and Power Blackouts}
\format{
A data frame with 12 observations on the following three variables:
\tabular{ll}{
\code{month} \tab month.\cr
\code{burglary} \tab number of burglaries.\cr
\code{blackout} \tab number of power blackouts.\cr
}
}
\source{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.burglary
}
\description{
This is a data set on the monthly number of burglaries and the number of power blackouts in a small town.
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 15) these hypothetical data are used to illustrate the consequences of a structural break.
}
\keyword{datasets}
