\name{data.cars}
\alias{data.cars}
\docType{data}
\title{
Speed of car and braking distance, used in ch. 5 of the workbook.
}
\description{
The data give the speed of cars and the distances taken to stop. Note that the data were recorded in the 1920s.
}
\usage{data.cars}
\format{
  A data frame of 50 observations with the following 2 variables:
  \describe{
    \item{country}{Speed (mph)}
    \item{income}{Stopping distance (ft)}
  }
}
\details{
These data provide a dataset used in chapter 5 of the workbook.
}
\source{
R package datasets
Ezekiel, M. (1930) Methods of Correlation Analysis. Wiley.
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
