% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.cobbdoug.R
\docType{data}
\name{data.cobbdoug}
\alias{data.cobbdoug}
\title{Cobb-Douglas Production Function}
\format{
A data frame with 100 observations on the following three variables:
\tabular{ll}{
\code{output} \tab production output.\cr
\code{labor} \tab input of labor.\cr
\code{capital} \tab input of capital.\cr
}
}
\source{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.cobbdoug
}
\description{
This is a data set on a production process that can be formalized by a Cobb-Douglas production function.
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 14) these hypothetical data are used to illustrate the functional specification of a regression model.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
