% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.comp.R
\docType{data}
\name{data.comp}
\alias{data.comp}
\title{Monthly Rentals and Qualitative Characteristics of Computers}
\format{
A data frame with 34 observations on the following four variables:
\tabular{ll}{
\code{rent} \tab monthly rental (in dollar).\cr
\code{mem} \tab memory capacity computed from three different computer characteristics.\cr
\code{access} \tab average time required to access information from memory.\cr
\code{mult} \tab average time required to obtain and complete multiplication instruction.\cr
}
}
\source{
Berndt, E.R. (1990): The Practice of Econometrics: Classic and Contemporary. Addison-Wesley, 136-142.
}
\usage{
data.comp
}
\description{
This is a data set on the monthly rentals of computers of different quality during the 1960s.
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chaps. 13 and 14) these hypothetical data are used to illustrate the specification of a multivariate regression model.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
