\name{data.fertilizer}
\alias{data.fertilizer}
\docType{data}
\title{
Fertilizer and barley crop yield (Ch. 8-12).
}
\description{
Hypothetical data of fertilizers (phospate and nitrogen) and barley output
}
\usage{data.fertilizer}
\format{
  A data frame with 30 observations on the following 3 variables.
  \describe{
    \item{phos}{Amount of phosphate.}
    \item{nit}{Amount of nitrogen.}
    \item{barley}{Barley crop yield.}
  }
}
\details{
These data are hypothetical and provide a dataset for estimating a multivariate linear model. Used in chapters 8-12 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
