\name{data.filter}
\alias{data.filter}
\docType{data}
\title{
Water filter sales data (Ch. 18).
}
\description{
Hypothetical water filter sales data: Monthly sales (in 1000 pcs.) and price (in Euro) of last 24 months.
}
\usage{data.filter}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{sales}{Monthly water filter sales (in 1000 pcs.).}
    \item{price}{price (in Euro)}
  }
}
\details{
These data provide a dataset for analyzing autocorrelation. Used in chapter 18 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
