\name{data.income}
\alias{data.income}
\docType{data}
\title{
Macroeconomic income data (Ch. 19).
}
\description{
Macroeconomic income data.
}
\usage{data.income}
\format{
  A data frame with 24 observations on the following 2 variables.
  \describe{
    \item{loginc}{Logarithmic per capita income.}
    \item{logsave}{Logarithmic savings rate.}
    \item{logsum}{Logarithmic sum of population growth rate, technical progress and capital depreciation.}
  }
}
\details{
These data provide a dataset for analyzing the assumption of normal errors. Used in chapter 19 of the textbook.
}
\source{
Mankiw, Romer and Weil (1992), "A Contribution to the Empirics of Economic Growth", Quarterly Journal of Economics, 107, 407-437.
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
