\name{data.insurance}
\alias{data.insurance}
\docType{data}
\title{
Insurance sales data (Ch. 20).
}
\description{
Hypothetical insurance sales data.
}
\usage{data.insurance}
\format{
  A data frame with 30 observations on the following 4 variables.
  \describe{
    \item{contr}{Number of insurance contracts currently sold.}
    \item{score}{Score of sales person in Assessment Center.}
    \item{contrprev}{Number of insurance contracts sold in previous period. Instrument variable}
    \item{realperf}{Real performance of sales person (not observable)}
  }
}
\details{
These data provide a dataset for demonstrating regression using random regressors and instrument variables. Used in chapter 20 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
