\name{data.iv}
\alias{data.iv}
\docType{data}
\title{
Instrument variable dataset.
}
\description{
Instrument variable dataset.
}
\usage{data.iv}
\format{
  To be added.
  \describe{
    \item{country}{Speed (mph)}
    \item{income}{Stopping distance (ft)}
  }
}
\details{
To be added.
}
\source{
To be added.
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
