\name{data.kmenta}
\alias{data.kmenta}
\docType{data}
\title{
BLAH
}
\description{
BLAH
}
\usage{data.kmenta}
\format{
  A data frame of 50 observations with the following 2 variables:
  \describe{
    \item{country}{Speed (mph)}
    \item{income}{Stopping distance (ft)}
  }
}
\details{
BLAH
}
\source{
R package datasets
Ezekiel, M. (1930) Methods of Correlation Analysis. Wiley.
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
