\name{data.lifesat}
\alias{data.lifesat}
\docType{data}
\title{
Another dataset on life satisfaction, used in ch. 3 of the workbook.
}
\description{
Life satisfaction and income data of 40 countries.
}
\usage{data.lifesat}
\format{
  A data frame of 40 observations with the following 2 variables.
  \describe{
    \item{country}{The country.}
    \item{income}{Per capita income in 1000 Dollars.}
    \item{lsat}{Life satisfaction index.}
  }
}
\details{
These data provide a dataset used in chapter 3 of the workbook.
}
\source{
World Value Survey
\url{http://www.worldvaluessurvey.org}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
