\name{data.milk}
\alias{data.milk}
\docType{data}
\title{
Milk production data (Ch. 14).
}
\description{
Hypothetical data of concentrated feed for cows and milk output.
}
\usage{data.milk}
\format{
  A data frame with 3 observations on the following 2 variables.
  \describe{
    \item{feed}{Concentrated feed (50kg/year).}
    \item{milk}{Milk output (litres/year.}
  }
}
\details{
These data are hypothetical and provide a minimal dataset for nonlinear relationships. Used in chapter 14 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
