\name{data.pharma}
\alias{data.pharma}
\docType{data}
\title{
Pharmaceutical advertisements data (Ch. 23).
}
\description{
Hypothetical quarterly commercials data of a pharmaceutical company.
}
\usage{data.pharma}
\format{
  A data frame with 24 observations on the following 4 variables.
  \describe{
    \item{sales}{Sales of pharmaceutical product (100g).}
    \item{ads}{Number of advertisements (double pages).}
    \item{price}{Price of pharmaceutical product (Euro/100g).}
    \item{adsprice}{Price of advertisements (1000 Euro/double page).}
  }
}
\details{
These data can be used to analyze simultaneous equation models. Used in chapter 23 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
