\name{data.printer}
\alias{data.printer}
\docType{data}
\title{
Technical data on laser printers (Ch. 21).
}
\description{
Technical data on laser printers from 1992-2001.
}
\usage{data.printer}
\format{
  A data frame with 44 observations on the following 5 variables.
  \describe{
    \item{price}{Price of the printer (Euro)0kg/year).}
    \item{speed}{Printer speed (pages per minute.}
    \item{size}{Printer size (decimeter^3.}
    \item{mcost}{Maintenance cost (Euro-Cent per page.}
    \item{tdiff}{Time difference to fist model considered (months).}
  }
}
\details{
These data can be used to analyze multicollinearity. Used in chapter 21 of the textbook.
}
\source{
Computer magazine C't
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
