\name{data.rent}
\alias{data.rent}
\docType{data}
\title{
Basic rent (without heating) of 12 districts (Ch. 17).
}
\description{
Distance to center and basic rent (without heating) of 12 districts.
}
\usage{data.rent}
\format{
  A data frame with 12 observations on the following 2 variables.
  \describe{
    \item{dist}{Distance to center (km).}
    \item{rent}{Basic rent (without heating) (Euro/m^2)}
  }
}
\details{
These data provide a dataset for analyzing heteroskedasticity. Used in chapter 17 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
