% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.sick.R
\docType{data}
\name{data.sick}
\alias{data.sick}
\title{Sick Leave and Unemployment}
\format{
A data frame with 23 observations on the following three variables:
\tabular{ll}{
\code{year} \tab year.\cr
\code{jobless} \tab average unemployment rate during that year (in percent).\cr
\code{sick} \tab average of employees' sick leave during that year (in percent).\cr
}
}
\source{
Federal Statistical Office of Germany (\url{https://www.destatis.de/DE/ZahlenFakten/Indikatoren/QualitaetArbeit/QualitaetDerArbeit.html}).
}
\usage{
data.sick
}
\description{
This is a data set on the unemployment rates and the sick leave in Germany in the years 1992 to 2014.
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 18) these data are used to illustrate the consequences of autocorrelated error terms.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
