\name{data.software}
\alias{data.software}
\docType{data}
\title{
Business data of a software company (Ch. 22).
}
\description{
Hypothetical monthly business data of a software company.
}
\usage{data.software}
\format{
  A data frame with 36 observations on the following 2 variables.
  \describe{
    \item{empl}{Employed people per month.}
    \item{orders}{Number of new orders.}
  }
}
\details{
These data can be used to analyze dynamical models. Used in chapter 22 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
