\name{data.theil}
\alias{data.theil}
\docType{data}
\title{
The title
}
\description{
The description
}
\usage{data.theil}
\format{
  A data frame with X observations on the following X variables.
  \describe{
    \item{Variable1}{Beschreibung.}
    \item{Variable2}{Beschreibung.}
  }
}
\details{
These data provide a dataset for analyzing XXX. Used in chapter XX of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
