\name{data.tip}
\alias{data.tip}
\docType{data}
\title{
Tip data in a restaurant (Ch. 2-7).
}
\description{
Hypothetical data of amount charged and tip in a restaurant
}
\usage{data.tip}
\format{
  A data frame with 3 observations on the following 2 variables.
  \describe{
    \item{x}{The amount charged on the check.}
    \item{y}{The tip given to the waitress.}
  }
}
\details{
These data are hypothetical and provide a minimal dataset for estimating a linear model. Used in chapters 2-7 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
