% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.trade.R
\docType{data}
\name{data.trade}
\alias{data.trade}
\title{Gravity Model Applied to Germany}
\format{
A data frame with 27 observations on the following five variables:
\tabular{ll}{
\code{country} \tab name of member state.\cr
\code{imports} \tab German imports from member state (in million euro).\cr
\code{exports} \tab German exports to member state (in million euro).\cr
\code{bip} \tab gross domestic product of member state (in million euro).\cr
\code{dist} \tab distance between member state and Germany (in km).\cr
}
}
\source{
Downloaded in 2015 from online platform of Eurostat (\url{https://ec.europa.eu/eurostat/home}).

Distances computed by FreeMapTools (\url{https://www.freemaptools.com}).
}
\usage{
data.trade
}
\description{
This is a data set on German trade with its 27 EU-partners in 2014.
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chaps. 9 and 14) these data are used to illustrate the estimation and functional specification of a multivariate linear regression model.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
