% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tsls.R
\docType{data}
\name{data.tsls}
\alias{data.tsls}
\title{Data TSLS}
\format{
A data frame with X observations on the following X variables.
\tabular{ll}{
\code{y} \tab endogenous variable.\cr
\code{x1} \tab first exogenous variable.\cr
\code{x2} \tab second exogenous variable.\cr
\code{z1} \tab first instrumental variable.\cr
\code{z2} \tab second instrumental variable.\cr
}
}
\source{
\url{https://www.oekonometrie-lernen.de}.
}
\usage{
data.tsls
}
\description{
The description
}
\details{
These data provide a dataset for analyzing XXX. Used in chapter XX of the textbook.
}
\references{
von Auer, L. (2016) Oekonometrie - Eine Einfuehrung, 7th Edition, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
