\name{data.unempl}
\alias{data.unempl}
\docType{data}
\title{
German unemployment rates (Ch. 15).
}
\description{
German economic growth and unemployment rates between 1992 and 2014
}
\usage{data.unempl}
\format{
  A data frame with 23 observations on the following 3 variables.
  \describe{
    \item{year}{The year.}
    \item{unempl}{Change in unemployment rate.}
    \item{gdp}{Change in Gross Domestic Product.}
  }
}
\details{
These data provide a dataset for analyzing structural changes. Used in chapter 15 of the textbook.
}
\source{
Federal Statistical Office Germany
\url{http://www.destatis.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
