\name{data.wage}
\alias{data.wage}
\docType{data}
\title{
Wage data in a company (Ch. 13 and 15).
}
\description{
Wage data in a company.
}
\usage{data.tip}
\format{
  A data frame with 3 observations on the following 2 variables.
  \describe{
    \item{wage}{Monthly wage in Euro.}
    \item{educ}{Education in years.}
    \item{age}{Age in years.}
    \item{empl}{Years employed.}
  }
}
\details{
These data are hypothetical and provide a minimal dataset for specification tests. Used in chapter 13 and 15 of the textbook.
}
\source{
\url{http://www.oekonometrie-lernen.de}
}
\references{
von Auer, L. (2016) 
\emph{Oekonometrie - Eine Einfuehrung}, 7th Edition, Springer-Gabler.
\url{http://www.oekonometrie-lernen.de}
}

\keyword{datasets}
