\docType{data}

\name{data.windscreen}

\alias{data.windscreen}

\title{Efficiency of a Car Glass Service Company}

\format{
A data frame with 248 observations on the following eight variables:
\tabular{ll}{
\code{screen} \tab number of windscreen replacements in the branch.\cr
\code{foreman} \tab foremen employed in the branch.\cr
\code{assist} \tab assistants employed in the branch.\cr
\code{f.wage} \tab foremen's average wage in the branch.\cr
\code{a.wage} \tab assistants' average wage in the branch.\cr
\code{f.age} \tab foremen's average age in the branch.\cr
\code{a.age} \tab assistants' average age in the branch.\cr
\code{capital} \tab total value of machines used for windscreen replacement in the branch (in euro).\cr
}
}

\source{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.windscreen
}

\description{
This is a data set on the business statistics of 248 branches of a car glass service company in 2015.
}

\details{
In Auer et al. (2023, Chap. 20) these hypothetical data illustrate the use of two stage least squares estimation with instrumental variables.
}

\keyword{datasets}
