\name{datasets}

\alias{datasets}

\title{(Datasets in desk.}

\description{
Generates a table of dataset names and descriptions available in package \code{desk}.
}

\usage{datasets()}

\arguments{
  None
}

\value{An object of class \code{table}}

\examples{
  datasets()
  }

\keyword{datasets, desk}