\name{f.coef.test}   % Der Name der Funktion

\alias{f.coef.test} 	% Kurzbeschreibung der Funktion

\title{F-test on multiple linear combinations of parameters in a lm model.}	% Titelzeile der Hilfe

\description{Performs an F-test (non-directional) on multiple (L) linear combinations of parameters in a linear model.}

\usage{
f.coef.test(mod, data = list(), nh, q = rep(0, dim(nh)[1]), sig.level = 0.05, details = FALSE, hyp = TRUE)
}

\arguments{
  \item{mod}{Model object estimated by \code{ols()} or \code{lm()}}
  
  \item{data}{Name of the dataframe to be used if \code{mod} is only a formula.}

  \item{nh}{Matrix of the coefficients of the linear combination of parameters. 
  Each of the L rows of that matrix represents a linear combination.}
  
  \item{q}{L-dimensional vector of values on which the parameter (combination) is to be tested against. Default value is the null-vector.}
  
  \item{sig.level}{Significance level. Default value: sig.level = 0.05.}
  
  \item{details}{Logical value indicating whether specific details about the test should be returned.}

  \item{hyp}{Logical value indicating whether the hypotheses shoild be part of the output. To be disabled if output is too large.}
}

\value{List including:

\tabular{ll}{
\code{hyp} \tab Character matrix of hypotheses (if hyp = TRUE).\cr
\code{results} \tab A dataframe of basic test results.\cr
\code{SSR.H0} \tab Sum of squared residuals in H0-model.\cr
\code{SSR.H1} \tab Sum of squared residuals in regular model.\cr
\code{nulldist} \tab Type of the Null distribution with its parameters.\cr
}

Objects x generated by f.coef.test can be plotted using plot(x, plot.what = "dist"). Argument "plot.what" can have values

\tabular{ll}{
\code{"dist"} \tab Plot the null distribution, test statistics and p-values.\cr
\code{"ellipse"} \tab Plot acceptance ellipse.\cr
}

If plot.what = "dist" is specified, futher arguments can be xpassed to plot():

\tabular{ll}{
\code{type = "acceptance"} \tab Plot acceptance ellipse ("acceptance") or confidence ellipse ("confidence")\cr
\code{which.coef = c(2,3)} \tab For which two coefficients should the ellipse be plotted?\cr
\code{center = TRUE} \tab Plot center of ellipse.\cr
\code{intervals = TRUE} \tab Plot interval borders.\cr
\code{test.point = TRUE} \tab Plot the point (q-values or coefficients) used in F-Test.\cr
\code{q = c(0,0)} \tab The q-value used in acceptance ellipse.\cr
\code{sig.level = 0.05} \tab Significance level used.\cr
}

}

\examples{
# H0: beta1 = 0.33 and beta2 = 0
x = f.coef.test(barley ~ phos + nit, data = log(data.fertilizer),
   nh = rbind(c(0,1,0), c(0,0,1)),
   q = c(0.33,0.33),
   details = TRUE)
x # Show the test results
plot(x) # Visualize the test result
plot(x, plot.what = "ellipse", q = c(0.33, 0.33))
}

\keyword{F-test, linear combination, linear hypothesis, linear model}