\name{ols.infocrit}

\alias{ols.infocrit}

\title{Calculate common information criteria.}

\description{
Calculates three common information criteria of models estimated by \code{ols()}.
}

\usage{ols.infocrit(mod, which = "all", scaled = FALSE)}

\arguments{
  \item{mod}{Linear model object generated by ols()}

  \item{which}{String value specifying the type of criterion: "aic" (Akaike Information Criterion), "sic" (Schwarz Information Criterion), or
  "pc", (Prognostic Criterion), optional, if ommitted then all criteria are returned ("all").}
  
  \item{scaled}{logical value which indicates whether criteria should be scaled by the number of observations T}
}

\value{Dataframe of AIC, SIC, and PC values}

\examples{
  y = c(4,1,3)
  x = c(1,2,4)
  my.mod = ols(y ~ x)
  ols.infocrit(my.mod) # Return all criteria unscaled
  ols.infocrit(my.mod, scaled = TRUE) # Return all criteria scaled
  ols.infocrit(my.mod, which = "pc") # Return Prognostic Criterion unscaled
}

\author{S. Hoffmann}

\keyword{Akaike Information Criterion, Schwarz Criterion, Prognostic Criterion}