\name{ols.predict}

\alias{ols.predict}

\title{Predictions in a linear model.}

\description{
Calculates the predicted values of a linear model based on specified values of the exogenous variables. Optionally the estimated variance of the prediction error is returned.
}

\usage{ols.predict(mod, data = list(), xnew, antilog = FALSE, details = FALSE)}

\arguments{
  \item{mod}{Model object generated by \code{ols()} or \code{lm()}}

  \item{data}{Name of dataframe to be specified if \code{mod} is a formula.}

  \item{xnew}{(T x K) matrix of new values of the exogenous variables, for which a prediction
    should be made, where
    \code{K} is the number of exogenous variables in the model
    \code{T} is the number of predictions to be made.
    If \code{xnew} is not specified, the fitted values are returned.}

  \item{antilog}{Logical value which indicates whether to re-transform the predicted value of a log transformed dependent variable back into original units.}

  \item{details}{Logical value, if specified as \code{TRUE}, a list is returned, which additionally includes the estimated variance of the prediction error (\code{var.pe}), estimated variance of the error term (\code{sig.squ}) and the estimated sampling error (\code{smpl.err})}
}

\value{A list object including:

\tabular{ll}{
\code{pred.val} \tab The predicted values.\cr
\code{xnew} \tab Values of predictor at which predictions should be evaluated.\cr
\code{var.pe} \tab Estimated variance of prediction error.\cr
\code{sig.squ} \tab Estimated variance of error term.\cr
\code{smpl.err} \tab Estimated sampling error.\cr
\code{mod} \tab The model estimated (for internal purposes)\cr
}
}

\examples{
  y = c(4,1,2,1)
  x1 = c(4,1,3,2)
  x2 = c(3,4,4,2)
  my.data = data.frame(y, x1, x2)

  # Returns fitted values
  ols.predict(y ~ x1 + x2, data = my.data)

  # Returns fitted values (equivalent to above)
  my.mod = ols(y ~ x1 + x2)
  ols.predict(my.mod)

  # Generate new x-data
  my.mat = rbind(c(x1 = 6, x2 = 5), c(x1 = 3, x2 = 3), c(x1 = 9, x2 = 10))
  
  # Returns predicted values
  ols.predict(my.mod, xnew = my.mat)

  # Returns predicted values and est. var. of pred. error
  ols.predict(my.mod, xnew = my.mat, details = TRUE)
}
\author{S. Hoffmann}
\keyword{predicted values, linear model}