\name{ols}   % Der Name der Funktion

\alias{ols} 	% Kurzbeschreibung der Funktion

\title{Ordinary Least Squares Regression}	% Titelzeile der Hilfe

\description{Estimates linear models using ordinary least squares estimation. Generated objects should be compatible with commands expecting objects generated by \code{lm()}}

\usage{ols(formula, data = list(), na.action = NULL, contrasts = NULL, details = FALSE, ...) }

\arguments{
  \item{formula}{Model formula.}

  \item{data}{Name of dataframe of variables in \code{formula}}

  \item{na.action}{Function which indicates what should happen when the data contain NAs.}

  \item{contrasts}{An optional list. See the contrasts.arg of \link[stats]{model.matrix.default}.}

  \item{details}{Logical value indicating whether details should be printed out by default.}

  \item{...}{Other arguments that \code{lm.fit()} supports.}
}

\value{List including:

\tabular{ll}{
\code{coefficients/coef} \tab Estimated parameters of the model.\cr
\code{residuals/resid} \tab Residuals of the estimation.\cr
\code{effects} \tab n vector of orthogonal single-df effects. The first rank of them correspond to non-aliased coefficients, and are named accordingly.\cr
\code{fitted.values} \tab Fitted values of the regression line.\cr
\code{df.residual/df} \tab Degrees of freedom in the model (Number of observations minus rank).\cr
\code{se} \tab Vector of standard errors of the parameter estimators.\cr
\code{t.value} \tab Vector of t-values of single parameter significance tests.\cr
\code{p.value} \tab Vector of p-values of single parameter significance tests.\cr
\code{data/model} \tab Matrix of the variables' data used.\cr
\code{response} \tab The endogenous (response) variable.\cr
\code{model.matrix} \tab The model (design) matrix.\cr
\code{ssr} \tab Sum of squared residuals.\cr
\code{sig.squ} \tab Estimated error variance (sigma squared).\cr
\code{vcov} \tab The variance-covariance matrix of the model's estimators.\cr
\code{r.squ} \tab Coefficient of determination (R squared).\cr
\code{adj.r.squ} \tab Adjusted coefficient of determination (adj. R squared).\cr
\code{nobs} \tab Number of observations.\cr
\code{ncoef/rank} \tab Integer, giving the rank of the model (Number of coefficients estimated)\cr
\code{has.const} \tab Logical value indicating whether model has constant parameter.\cr
\code{f.val} \tab F value for simultaneous significance of all slope parameters.\cr
\code{f.pval} \tab p value for simultaneous significance of all slope parameters.\cr
\code{modform} \tab The model's regression R-formula.\cr
\code{call} \tab The function call by which the regression was calcluated (Including \code{modform}).\cr
}
}

\examples{
# Minimal simple regression model
check <- c(10,30,50)
tip <- c(2,3,7)
tip.est <- ols (tip ~ check)

# Equivalent estimation using data argument
tip.est <- ols (y ~ x, data = data.tip)

# Show estimation results
tip.est 

# Show details
print(tip.est, details = TRUE)

# Plot scatter and regression line
plot(tip.est)

# Plot confidence (dark) and prediction bands (light), residuals and two center lines
plot(tip.est, pred.int = TRUE, conf.int = TRUE, residuals = TRUE, center = TRUE)

# Multiple regression model
fert.est <- ols(barley ~ phos + nit, data = log(data.fertilizer), details = TRUE)
fert.est
}

\keyword{ordinary least squares, linear model, regression}