\name{plot.desk}

\alias{plot.desk}

\title{Simplified plotting of regression- and test-results.}
\description{This function implements an S3 method for plotting regression- and test-results generated by functions of the desk package. Used for internal purposes.}

\usage{
\method{plot}{desk}(x, ...)
}

\arguments{
  \item{x}{Object of class desk to be plotted.} 
  \item{...}{Any argument that \code{plot()} accepts.}
}

\value{None}

\examples{
# Simple regression model
tip.est <- ols (y ~ x, data = data.tip)

# Check its class
class(tip.est)

# Standard regression plot (Scatter and regression line)
plot(tip.est)

# Regression plot with confidence- and prediction bands and residuals
plot(tip.est, pred.int = TRUE, conf.int = TRUE, resid = TRUE)
}

\keyword{plot, regression results, test results}