\name{print.desk}

\alias{print.desk}

\title{Alternative console output for regression- and test-results.}
\description{This function implements an S3 method for printing regression- and test-results generated by functions of the desk package. Used for internal purposes.}

\usage{
\method{print}{desk}(x, details, digits = 4, ...)
}

\arguments{
  \item{x}{Object of class desk to be printed to the console.} 
  \item{details}{Logical value indicating whether details should be printed.}
  \item{digits}{Number of digits to round to (only output).}
  \item{...}{Any argument that \code{print()} accepts.}
}

\value{None}

\examples{
# Simple regression model
tip.est <- ols (y ~ x, data = data.tip)

# Check its class
class(tip.est)

# Standard regression output
print(tip.est) # same as tip.est

# Regression output with details rounded to 2 digits
print(tip.est, details = TRUE, digits = 2)
}

\keyword{print, console, regression results, test results}