\name{rprofile.open}

\alias{rprofile.open}

\title{Open user R startup file Rprofile.site.}

\description{Opens the user R startup file "Rprofile.site" for viewing or editing.}

\usage{rprofile.open()}

\arguments{
None.
}

\value{None} % R?ckgabewert

\examples{
  \dontrun{rprofile.open()}
  }

\keyword{R, startup, configuration, Rprofile.site}