% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arguments.R
\name{arguments}
\alias{arguments}
\title{Arguments of a Function}
\usage{
arguments(fun, width = options("width")$width)
}
\arguments{
\item{fun}{name of the function.}

\item{width}{optional width for line breaking.}
}
\value{
None.
}
\description{
Shows the arguments and their default values of a function.
}
\examples{
arguments(sample.rep)

}
\seealso{
\code{\link[base]{args}}
}
\concept{default value}
\keyword{arguments,}
\keyword{default}
\keyword{function,}
\keyword{value}
