% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.anscombe.R
\docType{data}
\name{data.anscombe}
\alias{data.anscombe}
\title{Anscombe's Four Data Sets}
\format{
A data frame of 4 data sets, each with 11 observations of the two variables x and y.
\tabular{ll}{
\code{x1 to x4} \tab x-variables of the four data sets.\cr
\code{y1 to y4} \tab y-variables of the four data sets.\cr
}
}
\source{
Anscombe, F.J. (1973): Graphs in Statistical Analysis. American Statistician, 27(1), 17-21.
}
\usage{
data.anscombe
}
\description{
This data set comprises four individual x-y-data sets which have the same statistical properties (mean, variance, correlation, regression line, etc.), yet are quite different.
}
\details{
In von Auer and Hoffmann (2017, Chap. 3) these data are used to illustrate the simple regression model.
}
\references{
Tufte, E.R. (1989): The Visual Display of Quantitative Information, 13-14. Graphics Press.

Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
