\docType{data}

\name{data.ballb}

\alias{data.ballb}

\title{Defective Ball Bearings}

\format{
A data frame with six observations on the following two variables:
\tabular{ll}{
\code{defbb} \tab share of defective ball bearings (per thousand).\cr
\code{nshifts} \tab number of shifts between two maintenances.\cr
}
}

\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.ballb
}

\description{
This is a data set on the percentage of defective units in the production of ball bearings.
}

\details{
In Auer (2023, Chap. 16) and Auer et al. (2023, Chap. 16) these hypothetical data are used to illustrate the consequences of error terms with an expected value deviating from zero.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
