% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.ballb.R
\docType{data}
\name{data.ballb}
\alias{data.ballb}
\title{Defective Ball Bearings}
\format{
A data frame with seven observations on the following two variables:
\tabular{ll}{
\code{defbb} \tab share of defective ball bearings (per thousand).\cr
\code{nshifts} \tab number of shifts between two maintenances.\cr
}
}
\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.ballb
}
\description{
This is a data set on the percentage of rejects in the production of ball bearings.
}
\details{
In Auer (2023, Chap. 16) and Auer et al. (2023, Chap. 16) these hypothetical data are used to illustrate the consequences of error terms with an expected value deviating from zero.
}
\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
