% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.ballb.R
\docType{data}
\name{data.ballb}
\alias{data.ballb}
\title{Defective Ball Bearings}
\format{
A data frame with seven observations on the following two variables:
\tabular{ll}{
\code{defbb} \tab share of defective ball bearings (per thousand).\cr
\code{nshifts} \tab number of shifts between two maintenances.\cr
}
}
\source{
Auer, L. von (2016) Oekonometrie - Eine Einfuehrung, 7th ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.ballb
}
\description{
This is a data set on the percentage of rejects in the production of ball bearings.
}
\details{
In Auer, L. von (2016, Chap. 16) and Auer, L. von, S. Hoffmann (2017, Chap. 16) these hypothetical data are used to illustrate the consequences of error terms with an expected value deviating from zero.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
