\docType{data}

\name{data.burglary}

\alias{data.burglary}

\title{Burglaries and Power Blackouts}

\format{
A data frame with 12 observations on the following three variables:
\tabular{ll}{
\code{month} \tab month.\cr
\code{burglary} \tab number of burglaries.\cr
\code{blackout} \tab number of power blackouts.\cr
}
}

\source{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.burglary
}

\description{
This is a data set on the monthly number of burglaries and the number of power blackouts in a small town.
}

\details{
In Auer et al. (2023, Chap. 15) these hypothetical data are used to illustrate the consequences of a structural break.
}

\keyword{datasets}
