\docType{data}

\name{data.cars}

\alias{data.cars}

\title{Speed and Stopping Distances of Cars}

\format{
A data frame of 50 observations with the following two variables:
\tabular{ll}{
\code{speed} \tab speed (in miles per hour).\cr
\code{dist} \tab stopping distance (in foot).\cr
}
}

\source{
R package \pkg{datasets} (object \code{cars}).
Originally published in: Ezekiel, M. (1930): Methods of Correlation Analysis, Wiley.
}

\usage{
data.cars
}

\description{
The data give the speed of cars and the distances taken to stop. The data were recorded in the 1920s.
}

\details{
In Auer et al. (2023, Chaps. 5, 6, 7 & 16) the data are used to illustrate the simple regression model and the consequences of truncated data.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
