\docType{data}

\name{data.cobbdoug}

\alias{data.cobbdoug}

\title{Cobb-Douglas Production Function}

\format{
A data frame with 100 observations on the following three variables:
\tabular{ll}{
\code{output} \tab production output.\cr
\code{labor} \tab input of labor.\cr
\code{capital} \tab input of capital.\cr
}
}

\source{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.cobbdoug
}
\description{
This data set can be used to model a Cobb-Douglas production process.
}

\details{
In Auer et al. (2023, Chap. 14) these hypothetical data are used to illustrate the functional specification of a non-linear regression model.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
