\docType{data}

\name{data.fertilizer}

\alias{data.fertilizer}

\title{Fertilizer in the Cultivation of Barley}

\format{
A data frame with 30 observations on the following three variables:
\tabular{ll}{
\code{phos} \tab amount of phosphate (in kg per hectare).\cr
\code{nit} \tab amount of nitrogen (in kg per hectare).\cr
\code{barley} \tab barley crop yield (in units of 100 kg per hectare).\cr
}
}

\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.fertilizer
}

\description{
This is a data set on the use of fertilizers (phosphate and nitrogen) in the cultivation of barley.
}

\details{
In Auer (2023, Chap. 9) and Auer et al. (2023, Chap. 9). These hypothetical data are used to illustrate the estimation of a multivariate linear regression model.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
