% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.fertilizer.R
\docType{data}
\name{data.fertilizer}
\alias{data.fertilizer}
\title{Fertilizer in the Cultivation of Barley}
\format{
A data frame with 30 observations on the following three variables:
\tabular{ll}{
\code{phos} \tab amount of phosphate (in kg per hectare).\cr
\code{nit} \tab amount of nitrogen (in kg per hectare).\cr
\code{barley} \tab arley crop yield (in units of 100 kg per hectare).\cr
}
}
\source{
v.Auer, L. (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\usage{
data.fertilizer
}
\description{
This is a data set on the use of fertilizers (phospate and nitrogen) in the cultivation of barley.
}
\details{
In v.Auer, L. (2016, Chap. 9) and v.Auer, L. and Hoffmann, S. (2017, Chap. 9). These hypothetical data are used to illustrate the estimation of a multivariate linear regression model.
}
\references{
v.Auer, L. and Hoffmann, S. (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
