\docType{data}

\name{data.filter}

\alias{data.filter}

\title{Water Filter Sales}

\format{
A data frame with 24 observations on the following two variables:
\tabular{ll}{
\code{sales} \tab monthly water filter sales (in 1000 pcs.).\cr
\code{price} \tab price (in Euro).\cr
}
}

\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.filter
}

\description{
This is a data set on the prices and sales figures of water filters (in 1000 pcs.).
}

\details{
In Auer (2023, Chap. 18) and Auer et al. (2023, Chap. 18) these hypothetical data are used to illustrate the consequences of autocorrelated error terms.
}

\references{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).

Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
