% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.filter.R
\docType{data}
\name{data.filter}
\alias{data.filter}
\title{Water Filter Sales}
\format{
A data frame with 24 observations on the following two variables:
\tabular{ll}{
\code{sales} \tab monthly water filter sales (in 1000 pcs.).\cr
\code{price} \tab price (in Euro).\cr
}
}
\source{
Auer, L.v. (2016) Oekonometrie - Eine Einfuehrung, 7th ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.filter
}
\description{
This is a data set on the prices and sales figures of water filters (in 1000 pcs.).
}
\details{
In Auer, L.v. (2016, Chap. 18) and Auer, L.v., S.Hoffmann (2017, Chap. 18) these hypothetical data are used to illustrate the consequences of autocorrelated error terms.
}
\references{
Auer, L.v., S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
