% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.icecream.R
\docType{data}
\name{data.icecream}
\alias{data.icecream}
\title{Sales of Ice Cream}
\format{
A data frame with 35 observations on the following two variables:
\tabular{ll}{
\code{revenue} \tab revenues (in Euro).\cr
\code{temp} \tab temperature (in degree Celsius).\cr
}
}
\source{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.icecream
}
\description{
This data set is on the daily revenues from selling ice cream and the daily average temperature in some town on a sample of 35 working days.
}
\details{
In Auer et al. (2023, Chap. 7) these hypothetical data are used to illustrate the estimation of the simple linear regression model.
}
\keyword{datasets}
