% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.income.R
\docType{data}
\name{data.income}
\alias{data.income}
\title{Income Per Capita}
\format{
A data frame with 75 observations on the following three variables:
\tabular{ll}{
\code{loginc} \tab logarithmic per capita income.\cr
\code{logsave} \tab logarithmic savings rate.\cr
\code{logsum} \tab logarithmic sum of population growth rate, technical progress and capital depreciation.\cr
}
}
\source{
Mankiw, Romer and Weil (1992): A Contribution to the Empirics of Economic Growth. Quarterly Journal of Economics, 107, 407-437.
}
\usage{
data.income
}
\description{
This data set describes major macroeconomic variables determining the differences in per capita income of 75 countries in 1985.
}
\details{
In Auer, L von (2016, Chap. 19) and Auer, L. von, S. Hoffmann (2017, Chap. 19) these hypothetical data are used to illustrate the detection and consequences of error terms that are not normally distributed.
}
\references{
Auer, L. von (2016) Oekonometrie - Eine Einfuehrung, 7th ed., Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).

Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
