\docType{data}

\name{data.insurance}

\alias{data.insurance}

\title{Sales of Insurance Contracts}

\format{
A data frame with 30 observations on the following four variables:
\tabular{ll}{
\code{contr} \tab number of insurance contracts currently sold by the salesperson.\cr
\code{score} \tab score of salesperson in assessment center.\cr
\code{contrprev} \tab number of insurance contracts sold period by the salesperson in the previous.\cr
\code{ability} \tab salesperson's true ability to sell insurance contracts.\cr
}
}

\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.insurance
}

\description{
This is a data set on the ability and success of salespersons in selling insurance contracts.
}

\details{
In Auer (2023, Chap. 20) and Auer et al. (2023, Chap. 20) these hypothetical data illustrate the use of two stage least squares estimation with an instrumental variable.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
