% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.insurance.R
\docType{data}
\name{data.insurance}
\alias{data.insurance}
\title{Sales of Insurance Contracts}
\format{
A data frame with 30 observations on the following four variables:
\tabular{ll}{
\code{contr} \tab number of insurance contracts currently sold by the salesperson.\cr
\code{score} \tab score of salesperson in assessment center.\cr
\code{contrprev} \tab number of insurance contracts sold period by the salesperson in the previous.\cr
\code{ability} \tab salesperson's true ability to sell insurance contracts.\cr
}
}
\source{
Auer, L.v. (2016) Ökonometrie - Eine Einführung, 7th ed., Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\usage{
data.insurance
}
\description{
This is a data set on the ability and success of salespersons in selling insurance contracts.
}
\details{
In Auer (2016, Chap. 20) and Auer et al. (2017, Chap. 20) these hypothetical data illustrate the use of two stage least squares estimation with an instrumental variable.
}
\references{
Auer, L. von, S. Hoffmann (2017) Ökonometrie - Das R-Arbeitsbuch, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
