\docType{data}

\name{data.iv}

\alias{data.iv}

\title{Instrumental Variables}

\format{
A data frame with 8 observations on the following five variables:
\tabular{ll}{
\code{y} \tab endogenous variable.\cr
\code{x1} \tab first exogenous variable.\cr
\code{x2} \tab second exogenous variable.\cr
\code{z1} \tab first instrumental variable.\cr
\code{z2} \tab second instrumental variable.\cr
}
}

\source{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.iv
}

\description{
This data set is on the use of instrumental variables.
}

\details{
In Auer et al. (2023, Chap. 20) these hypothetical data are used to illustrate the use of two stage least squares estimation with instrumental variables.
}

\keyword{datasets}
