% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.iv.R
\docType{data}
\name{data.iv}
\alias{data.iv}
\title{Instrumental Variables}
\format{
A data frame with 8 observations on the following five variables:
\tabular{ll}{
\code{y} \tab endogenous variable.\cr
\code{x1} \tab first exogenous variable.\cr
\code{x2} \tab second exogenous variable.\cr
\code{z1} \tab first instrumental variable.\cr
\code{z2} \tab second instrumental variable.\cr
}
}
\source{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.iv
}
\description{
This data set is on the use of instrumental variables.
}
\details{
In Auer, L. von, S. Hoffmann (2017, Chap. 20) these hypothetical data are used to illustrate the use of two stage least squares estimation with instrumental variables.
}
\keyword{datasets}
