\docType{data}
\name{data.kmenta}
\alias{data.kmenta}
\title{Partly Artificial Data on the U. S. Economy}
\format{
A data frame of 20 observations with the following 5 variables:
\tabular{ll}{
\code{consump} \tab food consumption per capita.\cr
\code{price} \tab ratio of food prices to general consumer prices.\cr
\code{income} \tab disposable income in constant dollars.\cr
\code{farmPrice} \tab ratio of preceding year's prices received by farmers to general consumer prices.\cr
\code{trend} \tab time trend in years.\cr
}
}
\source{
R package \code{systemfit}.
}
\usage{
data.kmenta
}
\description{
These are partly contrived data from Kmenta (1986), constructed to illustrate estimation of a simultaneous-equation model.
}
\details{
The exogenous variables income, farmPrice, and trend are based on real data; the endogenous variables price and consump were generated by simulation.
}
\references{
Kmenta, J. (1986): Elements of Econometrics, 2nd Edition, Macmillan, New York.
}
\keyword{datasets}
