\docType{data}

\name{data.lifesat}

\alias{data.lifesat}

\title{Life Satisfaction}

\format{
A data frame of 40 observations with the following three variables:
\tabular{ll}{
\code{country} \tab country name.\cr
\code{income} \tab country's per capita income (in dollar).\cr
\code{lsat} \tab index of country's average life satisfaction.\cr
}
}

\source{
Imported from \href{https://www.worldvaluessurvey.org/WVSDocumentationWVL.jsp}{World Value Survey, Inglehart et al. (2014)}.
}

\usage{
data.lifesat
}

\description{
A data set describing the life satisfaction and per capita income in 40 countries in 2010.
}

\details{
In Auer et al. (2023, Chap. 3) these data are used to illustrate the use of the simple linear regression model.
}

\references{

Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).

Inglehart, R. et al. (2014): World Values Survey: All Rounds - Country-Pooled Datafile Version, R. Inglehart, C. Haerpfer, A. Moreno, C. Welzel, K. Kizilova, J. Diez-Medrano, M. Lagos, P. Norris, E. Ponarin & B. Puranen (eds.), Madrid: JD Systems Institute.
}

\keyword{datasets}
