\docType{data}

\name{data.milk}

\alias{data.milk}

\title{Milk Production}

\format{
A data frame with 12 observations on the following two variables:
\tabular{ll}{
\code{feed} \tab concentrated feed given to the cow (in units of 50kg per year).\cr
\code{milk} \tab milk output of the cow (in liters per year).\cr
}
}

\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.milk
}

\description{
This is a hypothetical data set on the use of concentrated feed for cows and their milk output.
}

\details{
In Auer (2023, Chap. 14) and Auer et al. (2023, Chap. 14) these hypothetical data are used to illustrate transformations in non-linear relationships.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
