% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.milk.R
\docType{data}
\name{data.milk}
\alias{data.milk}
\title{Milk Production}
\format{
A data frame with 12 observations on the following two variables:
\tabular{ll}{
\code{feed} \tab concentrated feed given to the cow (in units of 50kg per year).\cr
\code{milk} \tab milk output of the cow (in liters per year).\cr
}
}
\source{
Auer, L. von (2016) Oekonometrie - Eine Einfuehrung, 7th ed., Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\usage{
data.milk
}
\description{
This is a data set on the use of concentrated feed for cows and their milk output.
}
\details{
In Auer, L von (2016, Chap. 14) and Auer, L. von, S. Hoffmann (2017, Chap. 14) these hypothetical data are used to illustrate the functional specification of a regression model.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
