\docType{data}
\name{data.milk}
\alias{data.milk}
\title{Milk Production}
\format{
A data frame with 12 observations on the following two variables:
\tabular{ll}{
\code{feed} \tab concentrated feed given to the cow (in units of 50kg per year).\cr
\code{milk} \tab milk output of the cow (in liters per year).\cr
}
}
\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.milk
}
\description{
This is a data set on the use of concentrated feed for cows and their milk output.
}
\details{
In Auer (2023, Chap. 14) and Auer et al. (2023, Chap. 14) these hypothetical data are used to illustrate the functional specification of a regression model.
}
\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
