\docType{data}

\name{data.pharma}

\alias{data.pharma}

\title{Pharmaceutical Advertisements}

\format{
A data frame with 24 quarterly observations on the following four variables:
\tabular{ll}{
\code{sales} \tab sales of pharmaceutical product (in units of 100g).\cr
\code{ads} \tab number of advertisements (in double pages).\cr
\code{price} \tab price of pharmaceutical product (in euro per 100g).\cr
\code{adsprice} \tab price of advertisements (in units of 1000 euro per double page).\cr
}
}

\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.pharma
}

\description{
This is a data set on quarterly commercials data of a pharmaceutical company.
}

\details{
In Auer (2023, Chap. 23) and Auer et al. (2023, Chap. 23) these hypothetical data are used to illustrate the estimation of simultaneous equation econometric models.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
