% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.pharma.R
\docType{data}
\name{data.pharma}
\alias{data.pharma}
\title{Pharmaceutical Advertisements}
\format{
A data frame with 24 quarterly observations on the following four variables:
\tabular{ll}{
\code{sales} \tab sales of pharmaceutical product (in units of 100g).\cr
\code{ads} \tab number of advertisements (in double pages).\cr
\code{price} \tab price of pharmaceutical product (in euro per 100g).\cr
\code{adsprice} \tab price of advertisements (in units of 1000 euro per double page).\cr
}
}
\source{
Auer, L. von (2016) Oekonometrie - Eine Einfuehrung, 7th ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.pharma
}
\description{
This is a data set on quarterly commercials data of a pharmaceutical company.
}
\details{
In Auer, L von (2016, Chap. 23) and Auer, L. von, S. Hoffmann (2017, Chap. 23) these hypothetical data are used to illustrate the estimation of simultaneous equation econometric models.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
