\docType{data}

\name{data.printer}

\alias{data.printer}

\title{Prices and Qualitative Characteristics of Laser Printers}

\format{
A data frame with 44 observations on the following five variables:
\tabular{ll}{
\code{price} \tab price of the printer (in euro).\cr
\code{speed} \tab printer's speed (in pages per minute).\cr
\code{size} \tab printer's size (in cubic decimeter).\cr
\code{mcost} \tab maintenance costs of printer (in cent per page).\cr
\code{tdiff} \tab time difference between the printer's observation and the data set's first observed laser printer (in month).\cr
}
}

\source{
Data from computer magazin c't (February 1992 to August 2001).

Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.printer
}

\description{
This is a data set on the prices and qualitative characteristics of laser printers from 1992 to 2001.
}

\details{
In Auer (2023, Chap. 21) and Auer et al. (2023, Chap. 21) these hypothetical data are used to illustrate the consequences of multicollinear exogenous variables.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
