\docType{data}

\name{data.rent}

\alias{data.rent}

\title{Average Basic Rent in City Districts}

\format{
A data frame with 12 observations on the following four variables:
\tabular{ll}{
\code{rent} \tab district's basic rent (in euro per square meter).\cr
\code{dist} \tab distance between district and city center (in km).\cr
\code{share} \tab share of rental properties considered for random selection.\cr
\code{area} \tab usable area (in square meter).\cr
}
}

\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.rent
}

\description{
This is a hypothetical data set on twelve districts of a city. The data describe the district's distance to the city center and the average basic rent (it excludes additional costs).
}

\details{
In Auer (2023, Chap. 17) and Auer et al. (2023, Chap. 17) these hypothetical data are used to illustrate the consequences of heteroskedastic error terms.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de})..
}

\keyword{datasets}
