% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.rent.R
\docType{data}
\name{data.rent}
\alias{data.rent}
\title{Average Basic Rent in City Districts}
\format{
A data frame with 12 observations on the following four variables:
\tabular{ll}{
\code{rent} \tab district's basic rent (in euro per square meter).\cr
\code{dist} \tab distance between district and city center (in km).\cr
\code{share} \tab share of rental properties considered for random selection.\cr
\code{area} \tab usable area (in square meter).\cr
}
}
\source{
Auer, L. von (2016) Oekonometrie - Eine Einfuehrung, 7th ed., Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\usage{
data.rent
}
\description{
This is a data set on twelve districts of a city. The data describe the district's distance to the city center and the average basic rent (it excludes additional costs).
}
\details{
In Auer, L von (2016, Chap. 17) and Auer, L. von, S. Hoffmann (2017, Chap. 17) these hypothetical data are used to illustrate the consequences of heteroskedastic error terms.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
