\docType{data}

\name{data.software}

\alias{data.software}

\title{Employment Data of a Software Company}

\format{
A data frame with 36 observations on the following three variables:
\tabular{ll}{
\code{period} \tab identifies the time period. \cr
\code{empl} \tab number of employees in the observed month.\cr
\code{orders} \tab number of new orders during the observed month.\cr
}
}

\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.software
}

\description{
This is a hypothetical (time series) data set on business data of a software company covering 36 consecutive months.
}

\details{
In Auer (2023, Chap. 22) and Auer et al. (2023, Chap. 22) these hypothetical data are used to illustrate the estimation of dynamic regression models.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
