% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.software.R
\docType{data}
\name{data.software}
\alias{data.software}
\title{Employment Data of a Software Company}
\format{
A data frame with 36 observations on the following two variables:
\tabular{ll}{
\code{empl} \tab number of employees in the observed month.\cr
\code{orders} \tab number of new orders during the observed month.\cr
}
}
\source{
Auer, L. von (2016) Oekonometrie - Eine Einfuehrung, 7th ed., Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\usage{
data.software
}
\description{
This is a data set on business data of a software company covering 36 consecutive months.
}
\details{
In Auer, L von (2016, Chap. 22) and Auer, L. von, S. Hoffmann (2017, Chap. 22) these hypothetical data are used to illustrate the estimation of dynamic regression models.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
