% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.theil.R
\docType{data}
\name{data.theil}
\alias{data.theil}
\title{Data Theil}
\format{
A data frame with X observations on the following X variables.
\tabular{ll}{
\code{variable1} \tab Beschreibung.\cr
\code{variable2} \tab Beschreibung.\cr
}
}
\source{
\url{https://www.oekonometrie-lernen.de}.
}
\usage{
data.theil
}
\description{
The description
}
\details{
These data provide a dataset for analyzing XXX. Used in chapter XX of the textbook.
}
\references{
von Auer, L. (2016) Oekonometrie - Eine Einfuehrung, 7th Edition, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
