\docType{data}

\name{data.tip}

\alias{data.tip}

\title{Tip Data in a Restaurant}

\format{
A data frame with three observations on the following two variables:
\tabular{ll}{
\code{x} \tab the guest's bill (in euro).\cr
\code{y} \tab the tip given to the waiter/waitress (in euro).\cr
}
}

\source{
Auer, L.v. (2023): Ökonometrie - Eine Einführung, 8th ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\usage{
data.tip
}

\description{
This is a data set on the bills and the corresponding tips given in a restaurant of only 3 guests. Is can be used as minimal example to illustrate simple linear regression. The larger version of this dataset (20 guests) is available as \link{data.tip.all}.
}

\details{
In Auer (2023, Chap. 17) and Auer et al. (2023, Chap. 17) these hypothetical data provide a minimal data set for estimating a simple linear regression model.
}

\references{
Auer, L.v., Hoffmann, S. & Kranz, T. (2023): Ökonometrie - Das R-Arbeitsbuch, 2nd ed., Springer-Gabler (\href{https://www.uni-trier.de/index.php?id=15929}{https://www.oekonometrie-lernen.de}).
}

\keyword{datasets}
