% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.tip.R
\docType{data}
\name{data.tip}
\alias{data.tip}
\title{Tip Data in a Restaurant}
\format{
A data frame with three observations on the following two variables:
\tabular{ll}{
\code{x} \tab the guest's bill (in euro).\cr
\code{y} \tab the tip given to the waiter/waitress (in euro).\cr
}
}
\source{
v.Auer, L. (2016) \emph{Oekonometrie - Eine Einfuehrung}, 7th ed., Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\usage{
data.tip
}
\description{
This is a data set on the bills and the corresponding tips given in a restaurant.
}
\details{
In v.Auer, L. (2016, Chap. 17) and v.Auer, L. and Hoffmann, S. (2017, Chap. 17) these hypothetical data provide a minimal data set for estimating a simple linear regression model.
}
\references{
v.Auer, L. and Hoffmann, S. (2017) \emph{Oekonometrie - Das R-Arbeitsbuch}, Springer-Gabler (\bold{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
