% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.unempl.R
\docType{data}
\name{data.unempl}
\alias{data.unempl}
\title{German Economic Growth and Unemployment Rates}
\format{
A data frame with 30 observations on the following three variables:
\tabular{ll}{
\code{year} \tab year.\cr
\code{unempl} \tab change in German unemployment rate (in percentage points).\cr
\code{gdp} \tab change in German gross domestic product (in percentage).\cr
}
}
\source{
Online platform "Genesis" of the Federal Statistical Office of Germany (\url{https://www.destatis.de}).

Auer, L. von (2016) Oekonometrie - Eine Einfuehrung, 7th ed., Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\usage{
data.unempl
}
\description{
This is a data set on German economic growth and unemployment rates from 1992 to 2021.
}
\details{
In Auer, L von (2016, Chap. 15) and Auer, L. von, S. Hoffmann (2017, Chap. 15) these yearly data are used to illustrate the estimation of regression models that exhibit a structural break.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
