% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.wage.R
\docType{data}
\name{data.wage}
\alias{data.wage}
\title{Wage Data in a Company}
\format{
A data frame with 20 observations on the following four variables:
\tabular{ll}{
\code{wage} \tab employee's monthly wage (in euro).\cr
\code{educ} \tab employee's extra education beyond the basic schooling degree (in years).\cr
\code{age} \tab employee's age (in years).\cr
\code{empl} \tab employee's time of employment in the company (in years).\cr
}
}
\source{
Auer, L. von (2016) Oekonometrie - Eine Einfuehrung, 7th ed., Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\usage{
data.wage
}
\description{
This is a data set on the wage structure in a company.
}
\details{
In Auer, L von (2016, Chap. 13) and Auer, L. von, S. Hoffmann (2017, Chap. 13) these hypothetical data are used to illustrate the selection of the relevant exogenous variables.
}
\references{
Auer, L. von, S. Hoffmann (2017) Oekonometrie - Das R-Arbeitsbuch, Springer-Gabler (\url{https://www.oekonometrie-lernen.de}).
}
\keyword{datasets}
