% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{datasets}
\alias{datasets}
\title{Datasets in DESK}
\usage{
datasets()
}
\value{
An object of class \code{table}.
}
\description{
Generates a table of data set names and descriptions available in package \code{desk}.
}
\examples{
datasets()

}
\concept{datasets}
\concept{desk}
