% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.group.R
\name{get.group}
\alias{get.group}
\title{Get a Random Draw from Specified Group of Numbers}
\usage{
get.group(grps = seq(1, 5))
}
\arguments{
\item{grps}{vector of group numbers.}
}
\value{
None.
}
\description{
Get a random draw from specified group of numbers. This can be used for assigning groups of students to some exercises.
}
\examples{
get.group() # Draw group from group numbers 1 to 5
get.group(c(1,6,3)) # Draw group from specified group numbers

}
