% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivr.R
\name{ivr}
\alias{ivr}
\title{Two-Stage Least Squares (2SLS) Instrument Variable Regression}
\usage{
ivr(formula, data = list(), endog, iv, contrasts = NULL, details = FALSE, ...)
}
\arguments{
\item{formula}{model formula.}

\item{data}{name of the dataframe used. To be specified if variables are not stored in environment.}

\item{endog}{character vector of endogenous (to be instrumented) regressors.}

\item{iv}{character vector of (exogenous) instrumental variables.}

\item{contrasts}{an optional list. See the \code{contrasts.arg} of \code{\link[stats]{model.matrix.default}}.}

\item{details}{logical value indicating whether details should be printed out by default.}

\item{...}{further arguments that \code{lm.fit()} understands.}
}
\value{
A list object including:
\tabular{ll}{
\code{adj.r.squ} \tab adjusted coefficient of determination (adj. R-squared).\cr
\code{coefficients} \tab IV-estimators of model parameters.\cr
\code{data/model} \tab matrix of the variables' data used.\cr
\code{data.name} \tab name of the dataframe used.\cr
\code{df} \tab degrees of freedom in the model (number of observations minus rank).\cr
\code{exogenous} \tab exogenous regressors.\cr
\code{f.hausman} \tab exogeneity test: F-value for simultaneous significance of all instrument parameters. If H0: "Instruments are exogenous" is rejected, usage of IV-regression can be justified against OLS.\cr
\code{f.instr} \tab weak instrument test: F-value for significance of instrument parameter in first stage of 2SLS regression. If H0: "Instrument is weak" is rejected, instruments are usually considered sufficiently strong.\cr
\code{fitted.values} \tab fitted values of the IV-regression.\cr
\code{fsd} \tab first stage diagnostics (weakness of instruments).\cr
\code{has.const} \tab logical value indicating whether model has a constant (internal purposes).\cr
\code{instrumented} \tab name of instrumented regressors.\cr
\code{instruments} \tab name of instruments.\cr
\code{model.matrix} \tab the model (design) matrix.\cr
\code{ncoef} \tab integer, giving the rank of the model (number of coefficients estimated).\cr
\code{nobs} \tab number of observations.\cr
\code{p.hausman} \tab according p-value of exogeneity test.\cr
\code{p.instr} \tab according p-value of weak instruments test.\cr
\code{p.values} \tab vector of p-values of single parameter significance tests.\cr
\code{r.squ} \tab coefficient of determination (R-squared).\cr
\code{residuals} \tab residuals in the IV-regression.\cr
\code{response} \tab the endogenous (response) variable.\cr
\code{shea} \tab Shea's partial R-squared quantifying the ability to explain the endogenous regressors.\cr
\code{sig.squ} \tab estimated error variance (sigma-squared).\cr
\code{ssr} \tab sum of squared residuals.\cr
\code{std.err} \tab vector of standard errors of the parameter estimators.\cr
\code{t.values} \tab vector of t-values of single parameter significance tests.\cr
\code{ucov} \tab the (unscaled) variance-covariance matrix of the model's estimators.\cr
\code{vcov} \tab the (scaled) variance-covariance matrix of the model's estimators.\cr
\code{modform} \tab the model's regression R-formula.\cr
}
}
\description{
Performs a two-stage least squares regression.
}
\examples{
## Insurance Example, Chapter 20, Textbook
ivr(contr ~ score, endog = "score", iv = "contrprev", data = data.insurance, details = TRUE)

## Stata example
library(foreign)
data.educwages <- read.dta("http://www.stata-press.com/data/r12/educwages.dta")
ivr(wages ~ union + education, endog = "education", iv = c("meducation", "feducation"),
data = data.educwages)

## Replicating Murray (2006) Table 14.1 (page 612)
#data.graddy =
#read.table("http://wps.aw.com/wps/media/objects/2387/2445250/Data_Sets/ASCII/whiting.csv",
#header = TRUE, sep = ",")
#ivr(QTY ~ PRICE, endog = c("PRICE"), iv = "WINDSPD", details = TRUE, data = data.graddy)

## Replicating Murray (2006) Table 14.3 (page 622-623)
endog = c("PRICE")
#iv = c("WINDSPD","WINDSPD2", "STORMY", "COLD", "RAINY", "MIXED")
#ivr(QTY ~ PRICE + DAY1 + DAY2 + DAY3 + DAY4, endog = endog, iv = iv,
#details = TRUE, data = data.graddy)

endog = c("QTY")
#iv = iv = c("DAY1", "DAY2", "DAY3", "DAY4")
#ivr(PRICE ~ QTY + WINDSPD + WINDSPD2 + STORMY + COLD + RAINY + MIXED, endog = endog, iv = iv,
#details = TRUE, data = data.graddy)

## Demand and supply estimation (Kmenta, 1971)
## Copy variables to environment
consump <- data.kmenta$consump
price <- data.kmenta$price
income <- data.kmenta$income
farmPrice <- data.kmenta$farmPrice
trend <- data.kmenta$trend

## Define model equations
eqD <- consump ~ price + income
eqS <- consump ~ price + farmPrice + trend

## Estimate demand function with instruments "farmPrice"" and "trend""
D.est <- ivr(eqD, endog = "price", iv = c("farmPrice", "trend"))

## Estimate supply function with instrument "income"
S.est <- ivr(eqS, endog = "price", iv = "income")

## Calculate y-axis intersection and slope with averaged exogenous regressors
aD <- c(1, mean(income)) \%*\% D.est$coef[-2]
bD <- D.est$coef[2]
aS <- c(1, mean(farmPrice), mean(mean(trend))) \%*\% S.est$coef[-2]
bS <- S.est$coef[2]

## Plot data and estimated functions
plot(price, consump, xlim = c(0,120), ylim = c(0,120))
abline(a = aD, b = bD)
abline(a = aS, b = bS)

}
\references{
Greene, W.H. (1993): Econometric Analysis, 2nd ed., Macmillan.

Kmenta, J. (1971): Elements of Econometrics, 2nd ed., The University of Michigan Press.

Murray, M.P. (2006): Econometrics - A Modern Introduction, Pearson.
}
