% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ols.has.const.R
\name{ols.has.const}
\alias{ols.has.const}
\title{Check if Model has a Constant}
\usage{
ols.has.const(mod)
}
\arguments{
\item{mod}{linear model object of class "desk" or "lm".}
}
\value{
Logical value TRUE (has contant), FALSE (has no constant).
}
\description{
Checks if a linear model included a constant level parameter (alpha).
}
\examples{
my.modA = ols(y ~ x, data = data.tip)
my.modB = ols(y ~ 0 + x, data = data.tip)
ols.has.const(my.modA)
ols.has.const(my.modB)

}
