\name{plot.desk}
\alias{plot.desk}
\title{Simplified Plotting of Regression- and Test-results}
\usage{
\method{plot}{desk}(x, ...)
}
\arguments{
\item{x}{object of class desk to be plotted.}

\item{...}{any argument that \code{plot()} accepts.}
}
\description{
This function implements an S3 method for plotting regression- and test-results generated by functions of the desk package. Used for internal purposes.
}
\examples{
# see the help page of the commands, respectively

}
\concept{plot}
\concept{regression results}
\concept{test results}
