\name{rprofile.open}
\alias{rprofile.open}
\title{Open User R Startup File Rprofile.site}
\usage{
rprofile.open()
}
\value{
None.
}
\description{
Opens the user R startup file "Rprofile.site" for viewing or editing.
}
\examples{
if (FALSE) rprofile.open() # Open the file if statement = TRUE

}
\concept{R}
\concept{Rprofile.site}
\concept{configuration}
\concept{startup}
