% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showfile.R
\name{showfile}
\alias{showfile}
\title{Print ASCII File to Console}
\usage{
showfile(file, head = "all")
}
\arguments{
\item{file}{name of the file to be printed to the console.}

\item{head}{numeric value. Numbers of lines to be printed out. Default value is set to print everything.}
}
\value{
None.
}
\description{
Print an ASCII encoded text file to the console. Useful for quickly viewing .txt or .csv data sets.
}
\examples{
MyPath <- file.path(.libPaths()[1], "desk", "extdata", "Labor.txt")
showfile(MyPath) # show everything
showfile(MyPath, head = 4) # show only first 4 lines

}
\concept{arguments, function, default value}
